pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "legaldoc-ai"
version = "1.0.0"
description = "Multi-agent AI system for analyzing legal documents (NDAs)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "LegalDoc AI Team"}
]
keywords = ["legal", "nda", "ai", "llm", "document-analysis", "multi-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "instructor>=1.0.0",
    "tenacity>=8.0.0",
    "reportlab>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
legaldoc = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/legaldoc-ai"
Documentation = "https://github.com/yourusername/legaldoc-ai#readme"
Repository = "https://github.com/yourusername/legaldoc-ai"
Issues = "https://github.com/yourusername/legaldoc-ai/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["legaldoc*"]

[tool.setuptools.package-data]
"legaldoc.prompts" = ["*.txt"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["legaldoc"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src/legaldoc --cov-report=term-missing"
